requirements.txt
PyYAML==3.11
Markdown==2.5.2
python-dateutil==2.4.2
Django==1.8.18
Jinja2==2.8
MarkupSafe==1.0
django-extensions==1.7.9
django-mozilla-product-details==0.12.1
django-jinja==2.1.3
django-jinja-markdown==1.0.1
python-decouple==3.0
